fix(server-rust): drop unnecessary u32 cast on splice flags (Linux clippy)
`clippy -D warnings` failed on Linux at src/engine_raw.rs:630 and :666:
`(SPLICE_F_MOVE | SPLICE_F_MORE) as u32` is an unnecessary cast — the flags are
already `c_uint` (u32) and `libc::splice`'s flags param is `c_uint`. macOS doesn't
compile this Linux-only splice path, so local clippy didn't catch it. Verified the
Linux compile still builds (Docker).

+++ packages/server-rust/src/engine_raw.rs
@@ -627,7 +627,7 @@ pub(crate) mod zerocopy {
                                 pw,
                                 std::ptr::null_mut(),
                                 len,
-                                (libc::SPLICE_F_MOVE | libc::SPLICE_F_MORE) as u32,
+                                libc::SPLICE_F_MOVE | libc::SPLICE_F_MORE,
                             )
                         };
                         if n < 0 {
@@ -663,7 +663,7 @@ pub(crate) mod zerocopy {
                             file_fd,
                             file_off as *mut i64,
                             moved - drained,
-                            (libc::SPLICE_F_MOVE | libc::SPLICE_F_MORE) as u32,
+                            libc::SPLICE_F_MOVE | libc::SPLICE_F_MORE,
                         )
                     };
                     if m < 0 {
